Adding check for shadow and organizationtrail

+++ c7n/resources/cloudtrail.py
@@ -233,8 +233,13 @@ class DeleteTrail(BaseAction):
 
     def process(self, resources):
         client = local_session(self.manager.session_factory).client('cloudtrail')
+        shadow_check = IsShadow({'state': False}, self.manager)
+        shadow_check.embedded = True
+        resources = shadow_check.process(resources)
         for r in resources:
-            try:
-                client.delete_trail(Name=r['Name'])
-            except client.exceptions.TrailNotFoundException:
-                continue
+            if r.get('IsOrganizationTrail', False):
+                try:
+                    client.delete_trail(Name=r['Name'])
+                except client.exceptions.TrailNotFoundException:
+                    continue
+            continue
